+++ x/dex/keeper/grpc_query_test.go
@@ -243,7 +243,7 @@ func TestQueryNumPools(t *testing.T) {
 					/*poolId=*/ 1,
 					/*assets=*/ sdk.NewCoins(
 						sdk.NewInt64Coin("unibi", 100),
-						sdk.NewInt64Coin(common.DenomStable, 100),
+						sdk.NewInt64Coin(common.DenomNUSD, 100),
 					),
 					/*shares=*/ 100,
 				),
@@ -257,22 +257,22 @@ func TestQueryNumPools(t *testing.T) {
 					/*poolId=*/ 1,
 					/*assets=*/ sdk.NewCoins(
 						sdk.NewInt64Coin("unibi", 100),
-						sdk.NewInt64Coin(common.DenomColl, 100),
+						sdk.NewInt64Coin(common.DenomUSDC, 100),
 					),
 					/*shares=*/ 100,
 				),
 				mock.DexPool(
 					/*poolId=*/ 2,
 					/*assets=*/ sdk.NewCoins(
-						sdk.NewInt64Coin(common.DenomColl, 100),
-						sdk.NewInt64Coin(common.DenomStable, 100),
+						sdk.NewInt64Coin(common.DenomUSDC, 100),
+						sdk.NewInt64Coin(common.DenomNUSD, 100),
 					),
 					/*shares=*/ 100,
 				),
 				mock.DexPool(
 					/*poolId=*/ 3,
 					/*assets=*/ sdk.NewCoins(
-						sdk.NewInt64Coin(common.DenomStable, 100),
+						sdk.NewInt64Coin(common.DenomNUSD, 100),
 						sdk.NewInt64Coin("unibi", 100),
 					),
 					/*shares=*/ 100,
@@ -293,9 +293,9 @@ func TestQueryNumPools(t *testing.T) {
 				ctx,
 				sender,
 				sdk.NewCoins(
-					sdk.NewInt64Coin(common.DenomGov, 1e18),
-					sdk.NewInt64Coin(common.DenomStable, 1e18),
-					sdk.NewInt64Coin(common.DenomColl, 1e18),
+					sdk.NewInt64Coin(common.DenomNIBI, 1e18),
+					sdk.NewInt64Coin(common.DenomNUSD, 1e18),
+					sdk.NewInt64Coin(common.DenomUSDC, 1e18),
 				),
 			))
 
@@ -385,7 +385,7 @@ func TestQueryTotalShares(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
@@ -428,11 +428,11 @@ func TestQuerySpotPrice(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
-			tokenInDenom:  common.DenomStable,
+			tokenInDenom:  common.DenomNUSD,
 			tokenOutDenom: "unibi",
 			expectedPrice: sdk.MustNewDecFromStr("1"),
 		},
@@ -442,11 +442,11 @@ func TestQuerySpotPrice(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 200),
+					sdk.NewInt64Coin(common.DenomNUSD, 200),
 				),
 				/*shares=*/ 100,
 			),
-			tokenInDenom:  common.DenomStable,
+			tokenInDenom:  common.DenomNUSD,
 			tokenOutDenom: "unibi",
 			expectedPrice: sdk.MustNewDecFromStr("2"),
 		},
@@ -456,12 +456,12 @@ func TestQuerySpotPrice(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 200),
+					sdk.NewInt64Coin(common.DenomNUSD, 200),
 				),
 				/*shares=*/ 100,
 			),
 			tokenInDenom:  "unibi",
-			tokenOutDenom: common.DenomStable,
+			tokenOutDenom: common.DenomNUSD,
 			expectedPrice: sdk.MustNewDecFromStr("0.5"),
 		},
 	}
@@ -503,11 +503,11 @@ func TestQueryEstimateSwapExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
-			tokenIn:          sdk.NewInt64Coin(common.DenomStable, 100),
+			tokenIn:          sdk.NewInt64Coin(common.DenomNUSD, 100),
 			tokenOutDenom:    "unibi",
 			expectedTokenOut: sdk.NewInt64Coin("unibi", 50),
 		},
@@ -517,14 +517,14 @@ func TestQueryEstimateSwapExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 34844867),
-					sdk.NewInt64Coin(common.DenomStable, 4684496849),
+					sdk.NewInt64Coin(common.DenomNUSD, 4684496849),
 				),
 				/*shares=*/ 100,
 			),
 			tokenIn:       sdk.NewInt64Coin("unibi", 586848),
-			tokenOutDenom: common.DenomStable,
+			tokenOutDenom: common.DenomNUSD,
 			// https://www.wolframalpha.com/input?i=4684496849+-+%2834844867+*+4684496849+%2F+%2834844867%2B586848%29+%29
-			expectedTokenOut: sdk.NewInt64Coin(common.DenomStable, 77588330),
+			expectedTokenOut: sdk.NewInt64Coin(common.DenomNUSD, 77588330),
 		},
 	}
 
@@ -564,14 +564,14 @@ func TestQueryEstimateSwapExactAmountOut(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
 			tokenOut:     sdk.NewInt64Coin("unibi", 50),
-			tokenInDenom: common.DenomStable,
+			tokenInDenom: common.DenomNUSD,
 			// there's a swap fee that we take the ceiling of to round int
-			expectedTokenIn: sdk.NewInt64Coin(common.DenomStable, 101),
+			expectedTokenIn: sdk.NewInt64Coin(common.DenomNUSD, 101),
 		},
 		{
 			name: "complex swap",
@@ -579,11 +579,11 @@ func TestQueryEstimateSwapExactAmountOut(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 34844867),
-					sdk.NewInt64Coin(common.DenomStable, 4684496849),
+					sdk.NewInt64Coin(common.DenomNUSD, 4684496849),
 				),
 				/*shares=*/ 100,
 			),
-			tokenOut:     sdk.NewInt64Coin(common.DenomStable, 77588330),
+			tokenOut:     sdk.NewInt64Coin(common.DenomNUSD, 77588330),
 			tokenInDenom: "unibi",
 			// https://www.wolframalpha.com/input?i=4684496849+-+%2834844867+*+4684496849+%2F+%2834844867%2B586848%29+%29
 			expectedTokenIn: sdk.NewInt64Coin("unibi", 586848),
@@ -626,13 +626,13 @@ func TestQueryEstimateJoinExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
 			tokensIn: sdk.NewCoins(
 				sdk.NewInt64Coin("unibi", 100),
-				sdk.NewInt64Coin(common.DenomStable, 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
 			),
 			expectedPoolSharesOut: sdk.NewIntFromUint64(100),
 			expectedRemCoins:      sdk.NewCoins(),
@@ -643,17 +643,17 @@ func TestQueryEstimateJoinExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
 			tokensIn: sdk.NewCoins(
 				sdk.NewInt64Coin("unibi", 50),
-				sdk.NewInt64Coin(common.DenomStable, 75),
+				sdk.NewInt64Coin(common.DenomNUSD, 75),
 			),
 			expectedPoolSharesOut: sdk.NewIntFromUint64(50),
 			expectedRemCoins: sdk.NewCoins(
-				sdk.NewInt64Coin(common.DenomStable, 25),
+				sdk.NewInt64Coin(common.DenomNUSD, 25),
 			),
 		},
 	}
@@ -693,7 +693,7 @@ func TestQueryEstimateExitExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
@@ -701,7 +701,7 @@ func TestQueryEstimateExitExactAmountIn(t *testing.T) {
 			// exit fee leaves some tokens in pool
 			expectedTokensOut: sdk.NewCoins(
 				sdk.NewInt64Coin("unibi", 99),
-				sdk.NewInt64Coin(common.DenomStable, 99),
+				sdk.NewInt64Coin(common.DenomNUSD, 99),
 			),
 		},
 		{
@@ -710,7 +710,7 @@ func TestQueryEstimateExitExactAmountIn(t *testing.T) {
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),
-					sdk.NewInt64Coin(common.DenomStable, 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
 				),
 				/*shares=*/ 100,
 			),
@@ -718,7 +718,7 @@ func TestQueryEstimateExitExactAmountIn(t *testing.T) {
 			// exit fee leaves some tokens in pool
 			expectedTokensOut: sdk.NewCoins(
 				sdk.NewInt64Coin("unibi", 49),
-				sdk.NewInt64Coin(common.DenomStable, 49),
+				sdk.NewInt64Coin(common.DenomNUSD, 49),
 			),
 		},
 	}