+++ x/perp/keeper/msg_server_test.go
@@ -31,37 +31,37 @@ func TestMsgServerAddMargin(t *testing.T) {
 	}{
 		{
 			name:        "trader not enough funds",
-			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 999)),
+			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 999)),
 			initialPosition: &types.Position{
-				Pair:                           common.PairBTCStable,
+				Pair:                           common.Pair_BTC_NUSD,
 				Size_:                          sdk.OneDec(),
 				Margin:                         sdk.OneDec(),
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
 			},
-			margin:      sdk.NewInt64Coin(common.DenomStable, 1000),
+			margin:      sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr: sdkerrors.ErrInsufficientFunds,
 		},
 		{
 			name:            "no initial position",
-			traderFunds:     sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 1000)),
+			traderFunds:     sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1000)),
 			initialPosition: nil,
-			margin:          sdk.NewInt64Coin(common.DenomStable, 1000),
+			margin:          sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr:     types.ErrPositionNotFound,
 		},
 		{
 			name:        "success",
-			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 1000)),
+			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1000)),
 			initialPosition: &types.Position{
-				Pair:                           common.PairBTCStable,
+				Pair:                           common.Pair_BTC_NUSD,
 				Size_:                          sdk.OneDec(),
 				Margin:                         sdk.OneDec(),
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
 			},
-			margin:      sdk.NewInt64Coin(common.DenomStable, 1000),
+			margin:      sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr: nil,
 		},
 	}
@@ -76,7 +76,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 			t.Log("create vpool")
 			app.VpoolKeeper.CreatePool(
 				ctx,
-				common.PairBTCStable,
+				common.Pair_BTC_NUSD,
 				/* tradeLimitRatio */ sdk.OneDec(),
 				/* quoteReserve */ sdk.NewDec(1_000_000),
 				/* baseReserve */ sdk.NewDec(1_000_000),
@@ -86,7 +86,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
 			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
-				Pair:                   common.PairBTCStable,
+				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 
@@ -101,7 +101,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 
 			resp, err := msgServer.AddMargin(sdk.WrapSDKContext(ctx), &types.MsgAddMargin{
 				Sender:    traderAddr.String(),
-				TokenPair: common.PairBTCStable.String(),
+				TokenPair: common.Pair_BTC_NUSD.String(),
 				Margin:    tc.margin,
 			})
 
@@ -136,51 +136,51 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 	}{
 		{
 			name:       "position not enough margin",
-			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 1000)),
+			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1000)),
 			initialPosition: &types.Position{
-				Pair:                           common.PairBTCStable,
+				Pair:                           common.Pair_BTC_NUSD,
 				Size_:                          sdk.OneDec(),
 				Margin:                         sdk.OneDec(),
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
 			},
-			marginToRemove: sdk.NewInt64Coin(common.DenomStable, 1000),
+			marginToRemove: sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr:    types.ErrFailedRemoveMarginCanCauseBadDebt,
 		},
 		{
 			name:            "no initial position",
-			vaultFunds:      sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 0)),
+			vaultFunds:      sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 0)),
 			initialPosition: nil,
-			marginToRemove:  sdk.NewInt64Coin(common.DenomStable, 1000),
+			marginToRemove:  sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr:     types.ErrPositionNotFound,
 		},
 		{
 			name:       "vault insufficient funds",
-			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 999)),
+			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 999)),
 			initialPosition: &types.Position{
-				Pair:                           common.PairBTCStable,
+				Pair:                           common.Pair_BTC_NUSD,
 				Size_:                          sdk.OneDec(),
 				Margin:                         sdk.NewDec(1_000_000),
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
 			},
-			marginToRemove: sdk.NewInt64Coin(common.DenomStable, 1000),
+			marginToRemove: sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr:    sdkerrors.ErrInsufficientFunds,
 		},
 		{
 			name:       "success",
-			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 1000)),
+			vaultFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1000)),
 			initialPosition: &types.Position{
-				Pair:                           common.PairBTCStable,
+				Pair:                           common.Pair_BTC_NUSD,
 				Size_:                          sdk.OneDec(),
 				Margin:                         sdk.NewDec(1_000_000),
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
 			},
-			marginToRemove: sdk.NewInt64Coin(common.DenomStable, 1000),
+			marginToRemove: sdk.NewInt64Coin(common.DenomNUSD, 1000),
 			expectedErr:    nil,
 		},
 	}
@@ -195,7 +195,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 			t.Log("create vpool")
 			app.VpoolKeeper.CreatePool(
 				ctx,
-				common.PairBTCStable,
+				common.Pair_BTC_NUSD,
 				/* tradeLimitRatio */ sdk.OneDec(),
 				/* quoteReserve */ sdk.NewDec(1_000_000),
 				/* baseReserve */ sdk.NewDec(1_000_000),
@@ -205,7 +205,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
 			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
-				Pair:                   common.PairBTCStable,
+				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 
@@ -222,7 +222,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 
 			resp, err := msgServer.RemoveMargin(sdk.WrapSDKContext(ctx), &types.MsgRemoveMargin{
 				Sender:    traderAddr.String(),
-				TokenPair: common.PairBTCStable.String(),
+				TokenPair: common.Pair_BTC_NUSD.String(),
 				Margin:    tc.marginToRemove,
 			})
 
@@ -258,15 +258,15 @@ func TestMsgServerOpenPosition(t *testing.T) {
 	}{
 		{
 			name:        "trader not enough funds",
-			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 999)),
-			pair:        common.PairBTCStable.String(),
+			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 999)),
+			pair:        common.Pair_BTC_NUSD.String(),
 			sender:      sample.AccAddress().String(),
 			expectedErr: sdkerrors.ErrInsufficientFunds,
 		},
 		{
 			name:        "success",
-			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomStable, 1020)),
-			pair:        common.PairBTCStable.String(),
+			traderFunds: sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1020)),
+			pair:        common.Pair_BTC_NUSD.String(),
 			sender:      sample.AccAddress().String(),
 			expectedErr: nil,
 		},
@@ -281,7 +281,7 @@ func TestMsgServerOpenPosition(t *testing.T) {
 			t.Log("create vpool")
 			app.VpoolKeeper.CreatePool(
 				/* ctx */ ctx,
-				/* pair */ common.PairBTCStable,
+				/* pair */ common.Pair_BTC_NUSD,
 				/* tradeLimitRatio */ sdk.OneDec(),
 				/* quoteAssetReserve */ sdk.NewDec(1_000_000),
 				/* baseAssetReserve */ sdk.NewDec(1_000_000),
@@ -291,7 +291,7 @@ func TestMsgServerOpenPosition(t *testing.T) {
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
 			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
-				Pair:                   common.PairBTCStable,
+				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 
@@ -350,7 +350,7 @@ func TestMsgServerClosePosition(t *testing.T) {
 	}{
 		{
 			name:        "success",
-			pair:        common.PairBTCStable,
+			pair:        common.Pair_BTC_NUSD,
 			traderAddr:  sample.AccAddress(),
 			expectedErr: nil,
 		},
@@ -366,7 +366,7 @@ func TestMsgServerClosePosition(t *testing.T) {
 
 			app.VpoolKeeper.CreatePool(
 				ctx,
-				common.PairBTCStable,
+				common.Pair_BTC_NUSD,
 				/* tradeLimitRatio */ sdk.OneDec(),
 				/* quoteAssetReserve */ sdk.NewDec(1_000_000),
 				/* baseAssetReserve */ sdk.NewDec(1_000_000),
@@ -376,7 +376,7 @@ func TestMsgServerClosePosition(t *testing.T) {
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
 			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
-				Pair:                   common.PairBTCStable,
+				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 
@@ -433,21 +433,21 @@ func TestMsgServerLiquidate(t *testing.T) {
 		},
 		{
 			name:        "invalid liquidator address",
-			pair:        common.PairBTCStable.String(),
+			pair:        common.Pair_BTC_NUSD.String(),
 			liquidator:  "foo",
 			trader:      sample.AccAddress().String(),
 			expectedErr: fmt.Errorf("decoding bech32 failed"),
 		},
 		{
 			name:        "invalid trader address",
-			pair:        common.PairBTCStable.String(),
+			pair:        common.Pair_BTC_NUSD.String(),
 			liquidator:  sample.AccAddress().String(),
 			trader:      "foo",
 			expectedErr: fmt.Errorf("decoding bech32 failed"),
 		},
 		{
 			name:        "success",
-			pair:        common.PairBTCStable.String(),
+			pair:        common.Pair_BTC_NUSD.String(),
 			liquidator:  sample.AccAddress().String(),
 			trader:      sample.AccAddress().String(),
 			expectedErr: nil,
@@ -463,7 +463,7 @@ func TestMsgServerLiquidate(t *testing.T) {
 			t.Log("create vpool")
 			app.VpoolKeeper.CreatePool(
 				/* ctx */ ctx,
-				/* pair */ common.PairBTCStable,
+				/* pair */ common.Pair_BTC_NUSD,
 				/* tradeLimitRatio */ sdk.OneDec(),
 				/* quoteAssetReserve */ sdk.NewDec(1_000_000),
 				/* baseAssetReserve */ sdk.NewDec(1_000_000),
@@ -473,7 +473,7 @@ func TestMsgServerLiquidate(t *testing.T) {
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
 			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
-				Pair:                   common.PairBTCStable,
+				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 			ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 1).WithBlockTime(time.Now().Add(time.Minute))